requirements.txt
customtkinter==5.2.0
Pillow==10.0.0
CTkMessagebox==2.5
geocoder==1.38.1
folium==0.14.0
requests>=2.31.0
idna>=2.5
charset-normalizer~=2.0.0
urllib3>=1.21.1
certifi>=2017.4.17 